test: migrate misc traverse tests there is no more traverse proxy, but the test cases are still useful
