+++ packages/integration-karma/test/shadow-dom/Element-properties/Element.id.spec.js
@@ -1,0 +1,11 @@
+import { createElement } from 'test-utils';
+
+import TestWithDiv from 'x/testWithDiv';
+
+it('should allow setting id properties manually', () => {
+    const elm = createElement('x-foo', { is: TestWithDiv });
+    document.body.appendChild(elm);
+    const shadowRoot = elm.shadowRoot;
+    shadowRoot.querySelector('div').id = 'something';
+    expect(shadowRoot.querySelector('div').getAttribute('id')).toBe('something');
+});
