+++ packages/integration-karma/test/shadow-dom/Element-properties/Element.setAttribute.spec.js
@@ -1,0 +1,11 @@
+import { createElement } from 'test-utils';
+
+import TestWithDiv from 'x/testWithDiv';
+
+it('should allow setting attributes manually', () => {
+    const elm = createElement('x-foo', { is: TestWithDiv });
+    document.body.appendChild(elm);
+    const shadowRoot = elm.shadowRoot;
+    shadowRoot.querySelector('div').setAttribute('id', 'something');
+    expect(shadowRoot.querySelector('div').id).toBe('something');
+});
