+++ packages/integration-karma/test/shadow-dom/Element-properties/x/testWithDiv/testWithDiv.html
@@ -1,0 +1,3 @@
+<template>
+    <div></div>
+</template>
