+++ packages/integration-karma/test/shadow-dom/Element-properties/x/testWithDiv/testWithDiv.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class TestWithDiv extends LightningElement {}
